+++ packages/koot/defaults/before-build.js
@@ -3,6 +3,7 @@
  * - 文件名
  * - 代码中使用的临时常量名
  * - `koot.config.js` 的属性名
+ * - 其他代码中用到的常量名
  */
 module.exports = {
     // 处理项目配置相关
@@ -57,5 +58,9 @@ module.exports = {
     filenameSPATemplateInjectJS: '.spa-template-inject.LOCALEID.js',
 
     // [生产环境] 文件名
-    filenameCurrentBundle: '.koot-current'
+    filenameCurrentBundle: '.koot-current',
+
+    // CSS 相关
+    styleTagGlobalAttributeName: 'data-koot-global',
+    styleTagModuleAttributeName: 'data-koot-module'
 };