+++ packages/koot/index.js
@@ -7,6 +7,8 @@
 
 import isRenderSafe from './React/is-render-safe';
 
+// ============================================================================
+
 /**
  * 手动创建 Redux Store 时需要的相关对象
  * @type {Object}
@@ -22,7 +24,8 @@ export { createStore, reduxForCreateStore };
  */
 export { default as extend } from '__KOOT_HOC_EXTEND__';
 
-// 其他全局变量
+// ============================================================================
+
 export const getLocaleId = () => {
     if (__CLIENT__) return window.__KOOT_LOCALEID__ || '';
     if (__SERVER__) {
@@ -33,6 +36,8 @@ export const getLocaleId = () => {
 };
 export const resetLocaleId = () => (localeId = getLocaleId());
 export let localeId = (() => getLocaleId())();
+
+// ============================================================================
 
 export const getStore = () => {
     if (__CLIENT__) return window.__KOOT_STORE__;
@@ -45,6 +50,8 @@ export const getStore = () => {
 export const resetStore = () => (store = getStore());
 export let store = (() => getStore())();
 
+// ============================================================================
+
 export const getHistory = () => {
     if (__CLIENT__) return window.__KOOT_HISTORY__;
     if (__SERVER__) {
@@ -56,7 +63,7 @@ export const getHistory = () => {
 export const resetHistory = () => (history = getHistory());
 export let history = (() => getHistory())();
 
-//
+// ============================================================================
 
 export const getCache = localeId => {
     if (!isRenderSafe()) return {};
@@ -75,7 +82,7 @@ export const getCache = localeId => {
     }
 };
 
-//
+// ============================================================================
 
 if (__DEV__) {
     global.__KOOT_SSR_SET__ = v => {
@@ -93,4 +100,7 @@ if (__DEV__) {
         global.__KOOT_HISTORY__ = v;
         history = v;
     };
+    // if (__CLIENT__) {
+    //     window.__DEV_KOOT_GET_STYLES__ = getStyles;
+    // }
 }