+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 -   更新依赖包 minor 和 patch 版本号
 
+**koot**
+
+-   **新特性**
+    -   **新工具函数** `koot/utils/client-get-styles` - 获取当前全局 CSS 和所有组件 CSS。具体用法请参见文档 [全局与工具函数/工具函数/客户端](https://koot.js.org/#/utilities?id=客户端) ([#185](https://github.com/cmux/koot/issues/185))
+-   优化
+    -   组件 CSS 的 `<style>` 标签上不再有 `id` 属性，以避免和元素冲突
+
 ## [0.11.15] - 2019-10-15
 
 -   更新依赖包 minor 和 patch 版本号